backend/requirements.txt
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
chromadb==0.4.22
openai==1.13.3
anthropic==0.19.1
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.6
pydantic==2.6.3
beautifulsoup4==4.12.2
markdown==3.5.2
pytest==7.4.3
pytest-asyncio==0.23.2
tenacity==8.2.3
tiktoken==0.5.2
PyYAML==6.0.1
httpx==0.27.0
python-dotenv==1.0.1 